requirements.txt
streamlit==1.33.0
torch==2.2.2
torchvision==0.17.2
timm
pytorch-grad-cam
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.8.4
scikit-learn
opencv-python-headless
albumentations
pydicom
onnxruntime
